feat(as-3856): move site address e2e files to common and update imports
