+++ packages/e2e-tests/cypress/integration/householder-planning/appeals-service/appeal-submission/save-and-continue-navigation-appeal-site/save-and-continue-navigation-appeal-site.js
@@ -2,11 +2,11 @@ import { Given, When, Then } from 'cypress-cucumber-preprocessor/steps';
 import { answerDoesNotOwnTheWholeAppeal } from '../../../../../support/householder-planning/appeals-service/appeal-submission-appeal-site-ownership/answerDoesNotOwnTheWholeAppeal';
 import { clickSaveAndContinue } from '../../../../../support/householder-planning/appeals-service/appeal-navigation/clickSaveAndContinue';
 import { userIsNavigatedToPage } from '../../../../../support/householder-planning/appeals-service/appeal-navigation/userIsNavigatedToPage';
-import { provideAddressLine1 } from '../../../../../support/householder-planning/appeals-service/appeal-submission-appeal-site-address/provideAddressLine1';
-import { provideAddressLine2 } from '../../../../../support/householder-planning/appeals-service/appeal-submission-appeal-site-address/provideAddressLine2';
-import { provideTownOrCity } from '../../../../../support/householder-planning/appeals-service/appeal-submission-appeal-site-address/provideTownOrCity';
-import { provideCounty } from '../../../../../support/householder-planning/appeals-service/appeal-submission-appeal-site-address/provideCounty';
-import { providePostcode } from '../../../../../support/householder-planning/appeals-service/appeal-submission-appeal-site-address/providePostcode';
+import { provideAddressLine1 } from '../../../../../support/common/appeal-submission-appeal-site-address/provideAddressLine1';
+import { provideAddressLine2 } from '../../../../../support/common/appeal-submission-appeal-site-address/provideAddressLine2';
+import { provideTownOrCity } from '../../../../../support/common/appeal-submission-appeal-site-address/provideTownOrCity';
+import { provideCounty } from '../../../../../support/common/appeal-submission-appeal-site-address/provideCounty';
+import { providePostcode } from '../../../../../support/common/appeal-submission-appeal-site-address/providePostcode';
 import { answerOwnsTheWholeAppeal } from '../../../../../support/householder-planning/appeals-service/appeal-submission-appeal-site-ownership/answerOwnsTheWholeAppeal';
 import { answerHaveToldOtherOwnersAppeal } from '../../../../../support/householder-planning/appeals-service/appeal-submission-appeal-site-ownership/answerHaveToldOtherOwnersAppeal';
 import { answerHaveNotToldOtherOwnersAppeal } from '../../../../../support/householder-planning/appeals-service/appeal-submission-appeal-site-ownership/answerHaveNotToldOtherOwnersAppeal';
