+++ packages/e2e-tests/cypress/support/common/appeal-submission-appeal-site-address/confirmSiteAddressWasAccepted.js
