+++ packages/e2e-tests/cypress/support/householder-planning/appeals-service/back-link/navigateForwardsWithinAppealSteps.js
@@ -1,8 +1,8 @@
-import { provideAddressLine1 } from '../appeal-submission-appeal-site-address/provideAddressLine1';
-import { provideAddressLine2 } from '../appeal-submission-appeal-site-address/provideAddressLine2';
-import { provideTownOrCity } from '../appeal-submission-appeal-site-address/provideTownOrCity';
-import { provideCounty } from '../appeal-submission-appeal-site-address/provideCounty';
-import { providePostcode } from '../appeal-submission-appeal-site-address/providePostcode';
+import { provideAddressLine1 } from '../../../../support/common/appeal-submission-appeal-site-address/provideAddressLine1';
+import { provideAddressLine2 } from '../../../../support/common/appeal-submission-appeal-site-address/provideAddressLine2';
+import { provideTownOrCity } from '../../../../support/common/appeal-submission-appeal-site-address/provideTownOrCity';
+import { provideCounty } from '../../../../support/common/appeal-submission-appeal-site-address/provideCounty';
+import { providePostcode } from '../../../../support/common/appeal-submission-appeal-site-address/providePostcode';
 import { clickSaveAndContinue } from '../appeal-navigation/clickSaveAndContinue';
 import { answerOwnsTheWholeAppeal } from '../appeal-submission-appeal-site-ownership/answerOwnsTheWholeAppeal';
 import { answerCanSeeTheWholeAppeal } from '../appeal-submission-access-to-appeal-site/answerCanSeeTheWholeAppeal';
